Add multi-objective optimization solver for Pareto front
This script implements a multi-objective optimization solver using a classical approach to find the Pareto front of solutions based on two objectives: maximizing the number of 1s and counting flips in a binary string. It includes functions for defining objectives, solving the problem, filtering the Pareto front, and executing the main workflow.